Revert "Merge pull request #1601 from BoostryJP/feature/#1599"
This reverts commit 6e148b3d6b5d0ad96d30edbf03e9512657a3a1cf.
