+++ app/__init__.py
@@ -17,4 +17,4 @@ limitations under the License.
 SPDX-License-Identifier: Apache-2.0
 """
 
-__version__ = "25.6.0"
+__version__ = "25.3.0"
