+++ app/model/schema/__init__.py
@@ -120,8 +120,6 @@ from .token import (
     CreateTokenHoldersCollectionRequest,
     CreateTokenHoldersCollectionResponse,
     ListAllTokenHoldersQuery,
-    ListAllTokensQuery,
-    ListAllTokensResponse,
     ListAllTransferApprovalHistoryQuery,
     ListAllTransferHistoryQuery,
     RetrieveTokenHoldersCountQuery,
