+++ bin/healthcheck_indexer.sh
@@ -18,12 +18,11 @@
 # SPDX-License-Identifier: Apache-2.0
 
 if [[ "${APP_ENV:-}" != "local" && "${COMPANY_LIST_LOCAL_MODE:-}" -ne 1 ]]; then
-  PROC_LIST="${PROC_LIST} batch/indexer_Company_List.py"
+  PROC_LIST="${PROC_LIST} batch/indexer_CompanyList.py"
 fi
-PROC_LIST="${PROC_LIST} batch/indexer_Token_List.py"
+
 PROC_LIST="${PROC_LIST} batch/indexer_Transfer.py"
 PROC_LIST="${PROC_LIST} batch/indexer_Token_Holders.py"
-PROC_LIST="${PROC_LIST} batch/indexer_Token_List_Event.py"
 
 if [ "${SHARE_TOKEN_ENABLED}" = 1 ]; then
   PROC_LIST="${PROC_LIST} batch/indexer_Position_Share.py"
