+++ bin/run_indexer.sh
@@ -22,9 +22,8 @@ source ~/.bash_profile
 
 cd /app/ibet-Wallet-API
 
-
-# check COMPANY_LIST_LOCAL_MODE and COMPANY_LIST_URL
 if [[ "${APP_ENV:-}" != "local" && "${COMPANY_LIST_LOCAL_MODE:-}" -ne 1 ]]; then
+  # check COMPANY_LIST_LOCAL_MODE and COMPANY_LIST_URL
   if [ -z "${COMPANY_LIST_URL:-}" ]; then
     echo -n '[ERROR] Please set APP_ENV "local" or COMPANY_LIST_LOCAL_MODE "1", if you use company list local mode, ' >&2
     echo 'please set COMPANY_LIST_URL company list url, if you do not use local mode.' >&2
@@ -36,7 +35,7 @@ if [[ "${APP_ENV:-}" != "local" && "${COMPANY_LIST_LOCAL_MODE:-}" -ne 1 ]]; then
     echo -n "[WARNING] Could not access to COMPANY_LIST_URL, " >&2
     echo "please confirm COMPANY_LIST_URL, which response code is ${resp}" >&2
   fi
-  python batch/indexer_Company_List.py &
+  python batch/indexer_CompanyList.py &
 else
   # check company_list.json is default one
   content_length="$(wc -c data/company_list.json | awk '{print $1}')"
@@ -45,24 +44,9 @@ else
   fi
 fi
 
-
-# check TOKEN_LIST_URL
-if [ -z "${TOKEN_LIST_URL:-}" ]; then
-  echo -n '[ERROR] Please ensure that the environment variable TOKEN_LIST_URL is set.' >&2
-  exit 1
-fi
-# check TOKEN_LIST_URL available
-resp=$(curl "${TOKEN_LIST_URL}" -o /dev/null -w '%{http_code}\n' -s)
-if [ "${resp}" -ne 200 ]; then
-  echo -n "[WARNING] Could not access to TOKEN_LIST_URL, " >&2
-  echo "please confirm TOKEN_LIST_URL, which response code is ${resp}" >&2
-fi
-python batch/indexer_Token_List.py &
-
-
 python batch/indexer_Transfer.py &
 python batch/indexer_Token_Holders.py &
-python batch/indexer_Token_List_Event.py &
+python batch/indexer_Token_List.py &
 
 if [[ $SHARE_TOKEN_ENABLED = 1 ]]; then
   python batch/indexer_Position_Share.py &
