+++ tests/app/position_PositionMembership_test.py
@@ -26,7 +26,7 @@ from web3.middleware import ExtraDataToPOAMiddleware
 
 from app import config
 from app.contracts import Contract
-from app.model.db import IDXMembershipToken, IDXPosition, IDXTokenListRegister, Listing
+from app.model.db import IDXMembershipToken, IDXPosition, IDXTokenListItem, Listing
 from tests.account_config import eth_account
 from tests.contract_modules import (
     membership_issue,
@@ -174,7 +174,7 @@ class TestPositionMembership:
         idx_token.contact_information = "問い合わせ先"
         idx_token.privacy_policy = "プライバシーポリシー"
         session.add(idx_token)
-        idx_token_list_item = IDXTokenListRegister()
+        idx_token_list_item = IDXTokenListItem()
         idx_token_list_item.token_address = token_address
         idx_token_list_item.token_template = "IbetMembership"
         idx_token_list_item.owner_address = TestPositionMembership.issuer[
