+++ tests/app/token_share_ShareTokens_test.py
@@ -28,7 +28,7 @@ from web3.middleware import ExtraDataToPOAMiddleware
 
 from app import config
 from app.contracts import Contract
-from app.model.db import IDXTokenListRegister, Listing
+from app.model.db import IDXTokenListItem, Listing
 from batch import indexer_Token_Detail
 from batch.indexer_Token_Detail import Processor
 from tests.account_config import eth_account
@@ -100,7 +100,7 @@ class TestTokenShareTokens:
         listed_token.max_holding_quantity = 1
         listed_token.max_sell_amount = 1000
         session.add(listed_token)
-        token_list_item = IDXTokenListRegister()
+        token_list_item = IDXTokenListItem()
         token_list_item.token_address = token["address"]
         token_list_item.token_template = "IbetShare"
         token_list_item.owner_address = ""
