+++ tests/batch/indexer_Token_List_test.py
@@ -17,633 +17,883 @@ limitations under the License.
 SPDX-License-Identifier: Apache-2.0
 """
 
-import json
+import asyncio
 import logging
-from typing import Sequence
 from unittest import mock
-from unittest.mock import MagicMock
+from unittest.mock import AsyncMock, MagicMock
 
 import pytest
-import requests
 from sqlalchemy import select
+from sqlalchemy.exc import SQLAlchemyError
+from sqlalchemy.ext.asyncio import AsyncSession
+from sqlalchemy.orm import Session
+from web3 import Web3
+from web3.exceptions import ABIEventNotFound
+from web3.middleware import ExtraDataToPOAMiddleware
 
-from app.model.db import TokenList
-from batch.indexer_Token_List import LOG, Processor
+from app import config
+from app.contracts import Contract
+from app.errors import ServiceUnavailable
+from app.model.db import IDXTokenListBlockNumber, IDXTokenListItem, Listing
+from batch import indexer_Token_List
+from batch.indexer_Token_List import LOG, Processor, main
+from tests.account_config import eth_account
+from tests.contract_modules import (
+    coupon_register_list,
+    issue_bond_token,
+    issue_coupon_token,
+    issue_share_token,
+    membership_issue,
+    membership_register_list,
+    register_bond_list,
+    register_share_list,
+)
+
+web3 = Web3(Web3.HTTPProvider(config.WEB3_HTTP_PROVIDER))
+web3.middleware_onion.inject(ExtraDataToPOAMiddleware, layer=0)
+
+
+@pytest.fixture(scope="session")
+def test_module(shared_contract):
+    config.TOKEN_LIST_CONTRACT_ADDRESS = shared_contract["TokenList"]["address"]
+    return indexer_Token_List.Processor
 
 
 @pytest.fixture(scope="function")
-def processor(session):
-    return Processor()
+def processor(test_module, async_session):
+    LOG = logging.getLogger("ibet_wallet_batch")
+    default_log_level = LOG.level
+    LOG.setLevel(logging.DEBUG)
+    LOG.propagate = True
+
+    processor = test_module()
+    yield processor
+
+    LOG.propagate = False
+    LOG.setLevel(default_log_level)
 
 
 @pytest.fixture(scope="function")
-def caplog(caplog: pytest.LogCaptureFixture):
-    _log = logging.getLogger("ibet_wallet_batch")
-    default_log_level = _log.level
-    _log.setLevel(logging.INFO)
-    _log.propagate = True
-    yield caplog
-    _log.propagate = False
-    _log.setLevel(default_log_level)
-
-
-class MockResponse:
-    def __init__(self, data: object, status_code: int = 200):
-        self.data = data
-        self.status = status_code
-
-    async def __aexit__(self, exc_type, exc, tb):
-        pass
-
-    async def __aenter__(self):
-        return self
-
-    async def json(self) -> object:
-        return self.data
+def main_func(test_module):
+    LOG = logging.getLogger("ibet_wallet_batch")
+    default_log_level = LOG.level
+    LOG.setLevel(logging.DEBUG)
+    LOG.propagate = True
+    yield main
+    LOG.propagate = False
+    LOG.setLevel(default_log_level)
 
 
 @pytest.mark.asyncio
 class TestProcessor:
+    issuer = eth_account["issuer"]
+    user1 = eth_account["user1"]
+    user2 = eth_account["user2"]
+    trader = eth_account["trader"]
+    agent = eth_account["agent"]
+
+    @staticmethod
+    async def listing_token(token_address, async_session):
+        _listing = Listing()
+        _listing.token_address = token_address
+        _listing.is_public = True
+        _listing.max_holding_quantity = 1000000
+        _listing.max_sell_amount = 1000000
+        _listing.owner_address = TestProcessor.issuer["account_address"]
+        async_session.add(_listing)
+        await async_session.commit()
+
+    @staticmethod
+    def issue_token_bond_with_args(issuer, token_list, args):
+        # Issue token
+        token = issue_bond_token(issuer, args)
+        register_bond_list(issuer, token, token_list)
+
+        return token
+
+    @staticmethod
+    def issue_token_share_with_args(issuer, token_list, args):
+        # Issue token
+        token = issue_share_token(issuer, args)
+        register_share_list(issuer, token, token_list)
+
+        return token
+
+    @staticmethod
+    def issue_token_coupon_with_args(issuer, token_list, args):
+        # Issue token
+        token = issue_coupon_token(issuer, args)
+        coupon_register_list(issuer, token, token_list)
+
+        return token
+
+    @staticmethod
+    def issue_token_membership_with_args(issuer, token_list, args):
+        # Issue token
+        token = membership_issue(issuer, args)
+        membership_register_list(issuer, token, token_list)
+
+        return token
+
     ###########################################################################
     # Normal Case
     ###########################################################################
 
-    token_address_1 = "0xE883a6F441Ad5682D37Df31d34fC012bcb07A741"
-    token_address_2 = "0xE883A6f441AD5682D37df31d34FC012bcB07a742"
-    token_address_3 = "0xe883a6f441AD5682d37dF31D34fc012bCB07A743"
-    token_address_4 = "0xe883A6F441AD5682d37dF31d34fc012BcB07a744"
-    token_address_5 = "0xe883A6F441AD5682d37df31D34fc012BcB07A745"
-
     # <Normal_1>
-    # 0 record
-    @mock.patch("aiohttp.client.ClientSession.get")
-    async def test_normal_1(self, mock_get, processor, async_session):
-        # Prepare data
-        _token_list_item = TokenList()
-        _token_list_item.token_address = self.token_address_1
-        _token_list_item.token_template = "ibetBond"
-        _token_list_item.key_manager = ["0000000000000"]
-        _token_list_item.product_type = 1
-        async_session.add(_token_list_item)
-        _token_list_item = TokenList()
-        _token_list_item.token_address = self.token_address_2
-        _token_list_item.token_template = "ibetBond"
-        _token_list_item.key_manager = ["0000000000000"]
-        _token_list_item.product_type = 1
-        async_session.add(_token_list_item)
-        _token_list_item = TokenList()
-        _token_list_item.token_address = self.token_address_3
-        _token_list_item.token_template = "ibetShare"
-        _token_list_item.key_manager = ["1111111111111"]
-        _token_list_item.product_type = 5
-        async_session.add(_token_list_item)
+    # no token is listed
+    async def test_normal_1(
+        self,
+        processor: Processor,
+        shared_contract,
+        async_session: AsyncSession,
+        block_number: None,
+    ):
+        token_list_contract = shared_contract["TokenList"]
+        _token_list_block_number = IDXTokenListBlockNumber()
+        _token_list_block_number.latest_block_number = web3.eth.block_number
+        _token_list_block_number.contract_address = token_list_contract["address"]
+        async_session.add(_token_list_block_number)
         await async_session.commit()
-
-        # Mock
-        mock_get.side_effect = [MockResponse([])]
 
         # Run target process
         await processor.process()
 
-        # Assertion
-        await async_session.rollback()
-        _token_list: Sequence[TokenList] = (
-            await async_session.scalars(
-                select(TokenList).order_by(TokenList.token_address)
-            )
-        ).all()
+        # assertion
+        _token_list = (await async_session.scalars(select(IDXTokenListItem))).all()
         assert len(_token_list) == 0
 
     # <Normal_2>
-    # 1 record
-    @mock.patch("aiohttp.client.ClientSession.get")
-    async def test_normal_2(self, mock_get, processor, async_session):
-        # Prepare data
-        _token_list_item = TokenList()
-        _token_list_item.token_address = self.token_address_1
-        _token_list_item.token_template = "ibetBond"
-        _token_list_item.key_manager = ["0000000000000"]
-        _token_list_item.product_type = 1
-        async_session.add(_token_list_item)
-        _token_list_item = TokenList()
-        _token_list_item.token_address = self.token_address_2
-        _token_list_item.token_template = "ibetBond"
-        _token_list_item.key_manager = ["0000000000000"]
-        _token_list_item.product_type = 1
-        async_session.add(_token_list_item)
-        _token_list_item = TokenList()
-        _token_list_item.token_address = self.token_address_3
-        _token_list_item.token_template = "ibetShare"
-        _token_list_item.key_manager = ["1111111111111"]
-        _token_list_item.product_type = 5
-        async_session.add(_token_list_item)
+    # Multiple token is listed
+    async def test_normal_2(
+        self,
+        processor: Processor,
+        shared_contract,
+        async_session: AsyncSession,
+        block_number: None,
+    ):
+        token_list_contract = shared_contract["TokenList"]
+        personal_info_contract = shared_contract["PersonalInfo"]
+        exchange_contract = shared_contract["IbetStraightBondExchange"]
+
+        config.TOKEN_LIST_CONTRACT_ADDRESS = token_list_contract["address"]
+        _token_expected_list = []
+
+        _token_list_block_number = IDXTokenListBlockNumber()
+        _token_list_block_number.latest_block_number = web3.eth.block_number
+        _token_list_block_number.contract_address = token_list_contract["address"]
+        async_session.add(_token_list_block_number)
         await async_session.commit()
 
-        # Mock
-        mock_get.side_effect = [
-            MockResponse(
-                [
-                    {
-                        "token_address": self.token_address_4,
-                        "token_template": "ibetShare",
-                        "key_manager": ["0000000000000", "1111111111111"],
-                        "product_type": 1,
-                    },
-                ]
+        # Issue bond token
+        for i in range(2):
+            args = {
+                "name": f"テスト債券{str(i + 1)}",
+                "symbol": "BOND",
+                "totalSupply": 1000000,
+                "tradableExchange": exchange_contract["address"],
+                "faceValue": int((i + 1) * 10000),
+                "interestRate": 602,
+                "interestPaymentDate1": "0101",
+                "interestPaymentDate2": "0201",
+                "interestPaymentDate3": "0301",
+                "interestPaymentDate4": "0401",
+                "interestPaymentDate5": "0501",
+                "interestPaymentDate6": "0601",
+                "interestPaymentDate7": "0701",
+                "interestPaymentDate8": "0801",
+                "interestPaymentDate9": "0901",
+                "interestPaymentDate10": "1001",
+                "interestPaymentDate11": "1101",
+                "interestPaymentDate12": "1201",
+                "redemptionDate": "20191231",
+                "redemptionValue": 10000,
+                "returnDate": "20191231",
+                "returnAmount": "商品券をプレゼント",
+                "purpose": "新商品の開発資金として利用。",
+                "memo": "メモ",
+                "contactInformation": "問い合わせ先",
+                "privacyPolicy": "プライバシーポリシー",
+                "personalInfoAddress": personal_info_contract["address"],
+                "transferable": True,
+                "isRedeemed": False,
+                "faceValueCurrency": "JPY",
+                "interestPaymentCurrency": "JPY",
+                "redemptionValueCurrency": "JPY",
+                "baseFxRate": "",
+            }
+            token = self.issue_token_bond_with_args(
+                self.issuer, token_list_contract, args
             )
-        ]
-
-        # Run target process
-        await processor.process()
-
-        # Assertion
-        await async_session.rollback()
-        _token_list: Sequence[TokenList] = (
-            await async_session.scalars(
-                select(TokenList).order_by(TokenList.token_address)
+            await self.listing_token(token["address"], async_session)
+            _token_expected_list.append(
+                {
+                    "token_address": token["address"],
+                    "token_template": "IbetStraightBond",
+                    "owner_address": self.issuer["account_address"],
+                }
             )
-        ).all()
-        assert len(_token_list) == 1
-        assert _token_list[0].token_address == self.token_address_4  # checksum address
-        assert _token_list[0].token_template == "ibetShare"
-        assert _token_list[0].key_manager == ["0000000000000", "1111111111111"]
-        assert _token_list[0].product_type == 1
 
-    # <Normal_3>
-    # 2 record
-    @mock.patch("aiohttp.client.ClientSession.get")
-    async def test_normal_3(self, mock_get, processor, async_session):
-        # Prepare data
-        _token_list_item = TokenList()
-        _token_list_item.token_address = self.token_address_1
-        _token_list_item.token_template = "ibetBond"
-        _token_list_item.key_manager = ["0000000000000"]
-        _token_list_item.product_type = 1
-        async_session.add(_token_list_item)
-        _token_list_item = TokenList()
-        _token_list_item.token_address = self.token_address_2
-        _token_list_item.token_template = "ibetBond"
-        _token_list_item.key_manager = ["0000000000000"]
-        _token_list_item.product_type = 1
-        async_session.add(_token_list_item)
-        _token_list_item = TokenList()
-        _token_list_item.token_address = self.token_address_3
-        _token_list_item.token_template = "ibetShare"
-        _token_list_item.key_manager = ["1111111111111"]
-        _token_list_item.product_type = 5
-        async_session.add(_token_list_item)
-        await async_session.commit()
-
-        # Mock
-        mock_get.side_effect = [
-            MockResponse(
-                [
-                    {
-                        "token_address": self.token_address_4,
-                        "token_template": "ibetShare",
-                        "key_manager": ["0000000000000", "1111111111111"],
-                        "product_type": 1,
-                    },
-                    {
-                        "token_address": self.token_address_5,
-                        "token_template": "ibetBond",
-                        "key_manager": ["0000000000000", "1111111111111"],
-                        "product_type": 1,
-                    },
-                ]
+        # Issue share token
+        for i in range(2):
+            args = {
+                "name": "テスト株式",
+                "symbol": "SHARE",
+                "tradableExchange": exchange_contract["address"],
+                "personalInfoAddress": personal_info_contract["address"],
+                "issuePrice": int((i + 1) * 1000),
+                "principalValue": 1000,
+                "totalSupply": 1000000,
+                "dividends": 101,
+                "dividendRecordDate": "20200401",
+                "dividendPaymentDate": "20200502",
+                "cancellationDate": "20200603",
+                "contactInformation": "問い合わせ先",
+                "privacyPolicy": "プライバシーポリシー",
+                "memo": "メモ",
+                "transferable": True,
+            }
+            token = self.issue_token_share_with_args(
+                self.issuer, token_list_contract, args
             )
-        ]
-
-        # Run target process
-        await processor.process()
-
-        # Assertion
-        await async_session.rollback()
-        _token_list: Sequence[TokenList] = (
-            await async_session.scalars(
-                select(TokenList).order_by(TokenList.token_address)
+            await self.listing_token(token["address"], async_session)
+            _token_expected_list.append(
+                {
+                    "token_address": token["address"],
+                    "token_template": "IbetShare",
+                    "owner_address": self.issuer["account_address"],
+                }
             )
-        ).all()
-        assert len(_token_list) == 2
-        assert _token_list[0].token_address == self.token_address_4  # checksum address
-        assert _token_list[0].token_template == "ibetShare"
-        assert _token_list[0].key_manager == ["0000000000000", "1111111111111"]
-        assert _token_list[0].product_type == 1
-        assert _token_list[1].token_address == self.token_address_5  # checksum address
-        assert _token_list[1].token_template == "ibetBond"
-        assert _token_list[1].key_manager == ["0000000000000", "1111111111111"]
-        assert _token_list[1].product_type == 1
 
-    # <Normal_4_1>
-    # Insert SKIP
-    # Invalid type error
-    @mock.patch("aiohttp.client.ClientSession.get")
-    @pytest.mark.parametrize(
-        "invalid_record",
-        [
-            {
-                "token_address": "invalid_address",
-                "token_template": "ibetShare",
-                "key_manager": ["0000000000000", "1111111111111"],
-                "product_type": 1,
-            },  # invalid token_address
-            {
-                "token_address": token_address_5,
-                "token_template": 123,
-                "key_manager": ["0000000000000", "1111111111111"],
-                "product_type": 1,
-            },  # invalid token_template
-            {
-                "token_address": token_address_5,
-                "token_template": 123,
-                "key_manager": "0000000000000",
-                "product_type": 1,
-            },  # invalid key_manager
-            {
-                "token_address": token_address_5,
-                "token_template": 123,
-                "key_manager": "0000000000000",
-                "product_type": 1,
-            },  # invalid product_type
-            {
-                "token_template": "ibetShare",
-                "key_manager": ["0000000000000", "1111111111111"],
-                "product_type": 1,
-            },  # token_address missing
-            {
-                "token_address": token_address_5,
-                "key_manager": ["0000000000000", "1111111111111"],
-                "product_type": 1,
-            },  # token_template missing
-            {
-                "token_address": token_address_5,
-                "token_template": 123,
-                "product_type": 1,
-            },  # key_manager missing
-            {
-                "token_address": token_address_5,
-                "token_template": 123,
-                "key_manager": "0000000000000",
-            },  # product_type missing
-        ],
-    )
-    async def test_normal_4(self, mock_get, processor, async_session, invalid_record):
-        # Prepare data
-        _token_list_item = TokenList()
-        _token_list_item.token_address = self.token_address_1
-        _token_list_item.token_template = "ibetBond"
-        _token_list_item.key_manager = ["0000000000000"]
-        _token_list_item.product_type = 1
-        async_session.add(_token_list_item)
-        _token_list_item = TokenList()
-        _token_list_item.token_address = self.token_address_2
-        _token_list_item.token_template = "ibetBond"
-        _token_list_item.key_manager = ["0000000000000"]
-        _token_list_item.product_type = 1
-        async_session.add(_token_list_item)
-        _token_list_item = TokenList()
-        _token_list_item.token_address = self.token_address_3
-        _token_list_item.token_template = "ibetShare"
-        _token_list_item.key_manager = ["1111111111111"]
-        _token_list_item.product_type = 5
-        async_session.add(_token_list_item)
-        await async_session.commit()
-
-        # Mock
-        mock_get.side_effect = [
-            MockResponse(
-                [
-                    {
-                        "token_address": self.token_address_4,
-                        "token_template": "ibetShare",
-                        "key_manager": ["0000000000000", "1111111111111"],
-                        "product_type": 1,
-                    },
-                    invalid_record,
-                ]
+        # Issue membership token
+        for i in range(2):
+            args = {
+                "name": "テスト会員権",
+                "symbol": "MEMBERSHIP",
+                "initialSupply": int((i + 1) * 1000000),
+                "tradableExchange": exchange_contract["address"],
+                "details": "詳細",
+                "returnDetails": "リターン詳細",
+                "expirationDate": "20191231",
+                "memo": "メモ",
+                "transferable": True,
+                "contactInformation": "問い合わせ先",
+                "privacyPolicy": "プライバシーポリシー",
+            }
+            token = self.issue_token_membership_with_args(
+                self.issuer, token_list_contract, args
             )
-        ]
-
-        # Run target process
-        await processor.process()
-
-        # Assertion
-        await async_session.rollback()
-        _token_list: Sequence[TokenList] = (
-            await async_session.scalars(
-                select(TokenList).order_by(TokenList.token_address)
+            await self.listing_token(token["address"], async_session)
+            _token_expected_list.append(
+                {
+                    "token_address": token["address"],
+                    "token_template": "IbetMembership",
+                    "owner_address": self.issuer["account_address"],
+                }
             )
-        ).all()
-        assert len(_token_list) == 1
-        assert _token_list[0].token_address == self.token_address_4  # checksum address
-        assert _token_list[0].token_template == "ibetShare"
-        assert _token_list[0].key_manager == ["0000000000000", "1111111111111"]
-        assert _token_list[0].product_type == 1
 
-    # <Normal_5_1>
-    # There are no differences from last time
-    # -> Skip this cycle
-    @mock.patch("aiohttp.client.ClientSession.get")
-    async def test_normal_5_1(self, mock_get, processor, async_session, caplog):
-        # Run target process: 1st time
-        mock_get.side_effect = [
-            MockResponse(
-                [
-                    {
-                        "token_address": self.token_address_4,
-                        "token_template": "ibetShare",
-                        "key_manager": ["0000000000000", "1111111111111"],
-                        "product_type": 1,
-                    },
-                    {
-                        "token_address": self.token_address_5,
-                        "token_template": "ibetBond",
-                        "key_manager": ["0000000000000", "1111111111111"],
-                        "product_type": 1,
-                    },
-                ]
-            )
-        ]
-        await processor.process()
+        # Issue coupon token
+        for i in range(2):
+            args = {
+                "name": "テストクーポン",
+                "symbol": "COUPON",
+                "totalSupply": int((i + 1) * 1000000),
+                "tradableExchange": exchange_contract["address"],
+                "details": "クーポン詳細",
+                "returnDetails": "リターン詳細",
+                "memo": "クーポンメモ欄",
+                "expirationDate": "20191231",
+                "transferable": True,
+                "contactInformation": "問い合わせ先",
+                "privacyPolicy": "プライバシーポリシー",
+            }
 
-        # Run target process: 2nd time
-        mock_get.side_effect = [
-            MockResponse(
-                [
-                    {
-                        "token_address": self.token_address_4,
-                        "token_template": "ibetShare",
-                        "key_manager": ["0000000000000", "1111111111111"],
-                        "product_type": 1,
-                    },
-                    {
-                        "token_address": self.token_address_5,
-                        "token_template": "ibetBond",
-                        "key_manager": ["0000000000000", "1111111111111"],
-                        "product_type": 1,
-                    },
-                ]
+            token = self.issue_token_coupon_with_args(
+                self.issuer, token_list_contract, args
             )
-        ]
-        await processor.process()
+            await self.listing_token(token["address"], async_session)
+            _token_expected_list.append(
+                {
+                    "token_address": token["address"],
+                    "token_template": "IbetCoupon",
+                    "owner_address": self.issuer["account_address"],
+                }
+            )
 
-        # Assertion
-        await async_session.rollback()
-        _token_list: Sequence[TokenList] = (
-            await async_session.scalars(
-                select(TokenList).order_by(TokenList.token_address)
-            )
-        ).all()
-        assert len(_token_list) == 2
-        assert _token_list[0].token_address == self.token_address_4  # checksum address
-        assert _token_list[0].token_template == "ibetShare"
-        assert _token_list[0].key_manager == ["0000000000000", "1111111111111"]
-        assert _token_list[0].product_type == 1
-        assert _token_list[1].token_address == self.token_address_5  # checksum address
-        assert _token_list[1].token_template == "ibetBond"
-        assert _token_list[1].key_manager == ["0000000000000", "1111111111111"]
-        assert _token_list[1].product_type == 1
-
-        assert 1 == caplog.record_tuples.count(
-            (
-                LOG.name,
-                logging.INFO,
-                "Skip: There are no differences from the previous cycle",
-            )
+        # register unknown token template
+        TokenListContract = Contract.get_contract(
+            "TokenList", token_list_contract["address"]
         )
+        web3.eth.default_account = self.issuer["account_address"]
+        args = {
+            "name": "TestToken",
+            "symbol": "Test",
+            "totalSupply": 1000000,
+            "tradableExchange": exchange_contract["address"],
+            "faceValue": 10000,
+            "interestRate": 602,
+            "interestPaymentDate1": "0101",
+            "interestPaymentDate2": "0201",
+            "interestPaymentDate3": "0301",
+            "interestPaymentDate4": "0401",
+            "interestPaymentDate5": "0501",
+            "interestPaymentDate6": "0601",
+            "interestPaymentDate7": "0701",
+            "interestPaymentDate8": "0801",
+            "interestPaymentDate9": "0901",
+            "interestPaymentDate10": "1001",
+            "interestPaymentDate11": "1101",
+            "interestPaymentDate12": "1201",
+            "redemptionDate": "20191231",
+            "redemptionValue": 10000,
+            "returnDate": "20191231",
+            "returnAmount": "商品券をプレゼント",
+            "purpose": "新商品の開発資金として利用。",
+            "memo": "メモ",
+            "contactInformation": "問い合わせ先",
+            "privacyPolicy": "プライバシーポリシー",
+            "personalInfoAddress": personal_info_contract["address"],
+            "transferable": True,
+            "isRedeemed": False,
+            "faceValueCurrency": "JPY",
+            "interestPaymentCurrency": "JPY",
+            "redemptionValueCurrency": "JPY",
+            "baseFxRate": "",
+        }
+        test_token = issue_bond_token(self.issuer, args)
+        TokenListContract.functions.register(
+            test_token["address"], "UnknownTokenTemplate"
+        ).transact({"from": self.issuer["account_address"]})
 
-    # <Normal_5_2>
-    # There are differences from the previous cycle
-    @mock.patch("aiohttp.client.ClientSession.get")
-    async def test_normal_5_2(self, mock_get, processor, async_session, caplog):
-        # Run target process: 1st time
-        mock_get.side_effect = [
-            MockResponse(
-                [
-                    {
-                        "token_address": self.token_address_4,
-                        "token_template": "ibetShare",
-                        "key_manager": ["0000000000000", "1111111111111"],
-                        "product_type": 1,
-                    },
-                    {
-                        "token_address": self.token_address_5,
-                        "token_template": "ibetBond",
-                        "key_manager": ["0000000000000", "1111111111111"],
-                        "product_type": 1,
-                    },
-                ]
-            )
-        ]
+        # Run target process
         await processor.process()
 
-        # Run target process: 2nd time
-        mock_get.side_effect = [
-            MockResponse(
-                [
-                    {
-                        "token_address": self.token_address_4,
-                        "token_template": "ibetShare",
-                        "key_manager": ["0000000000000", "1111111111111"],
-                        "product_type": 1,
-                    },
-                ]
-            )
-        ]
-        await processor.process()
+        # assertion
+        for _expect_dict in _token_expected_list:
+            token_list_item: IDXTokenListItem = (
+                await async_session.scalars(
+                    select(IDXTokenListItem)
+                    .where(
+                        IDXTokenListItem.token_address == _expect_dict["token_address"]
+                    )
+                    .limit(1)
+                )
+            ).first()
 
-        # Assertion
-        await async_session.rollback()
-        _token_list: Sequence[TokenList] = (
-            await async_session.scalars(
-                select(TokenList).order_by(TokenList.token_address)
-            )
-        ).all()
-        assert len(_token_list) == 1
-        assert _token_list[0].token_address == self.token_address_4  # checksum address
-        assert _token_list[0].token_template == "ibetShare"
-        assert _token_list[0].key_manager == ["0000000000000", "1111111111111"]
-        assert _token_list[0].product_type == 1
+            assert token_list_item.token_template == _expect_dict["token_template"]
+            assert token_list_item.owner_address == _expect_dict["owner_address"]
 
-        assert 0 == caplog.record_tuples.count(
-            (
-                LOG.name,
-                logging.INFO,
-                "Skip: There are no differences from the previous cycle",
-            )
-        )
+    # <Normal_3_1>
+    # When processed block_number is not stored and current block number is 9,999,999,
+    # then processor should call "__sync_register" method 10 times.
+    async def test_normal_3_1(
+        self,
+        processor: Processor,
+        shared_contract,
+        async_session: AsyncSession,
+        block_number: None,
+        caplog: pytest.LogCaptureFixture,
+    ):
+        token_list_contract = shared_contract["TokenList"]
+
+        config.TOKEN_LIST_CONTRACT_ADDRESS = token_list_contract["address"]
+
+        mock_lib = MagicMock()
+
+        current_block_number = 10000000 - 1
+        block_number_mock = AsyncMock()
+        block_number_mock.return_value = current_block_number
+
+        # Run target process
+        with mock.patch(
+            "web3.eth.async_eth.AsyncEth.block_number", block_number_mock()
+        ):
+            with mock.patch.object(
+                Processor, "_Processor__sync_register", return_value=mock_lib
+            ) as __sync_register_mock:
+                __sync_register_mock.return_value = None
+                await processor.process()
+
+                # Then processor calls "__sync_register" method 10 times.
+                assert __sync_register_mock.call_count == 10
+
+                idx_token_list_block_number: IDXTokenListBlockNumber = (
+                    await async_session.scalars(
+                        select(IDXTokenListBlockNumber).limit(1)
+                    )
+                ).first()
+                assert (
+                    idx_token_list_block_number.latest_block_number
+                    == current_block_number
+                )
+
+                assert 1 == caplog.record_tuples.count(
+                    (LOG.name, logging.INFO, "Syncing from=0, to=999999")
+                )
+                assert 1 == caplog.record_tuples.count(
+                    (LOG.name, logging.INFO, "Syncing from=1000000, to=1999999")
+                )
+                assert 1 == caplog.record_tuples.count(
+                    (LOG.name, logging.INFO, "Syncing from=2000000, to=2999999")
+                )
+                assert 1 == caplog.record_tuples.count(
+                    (LOG.name, logging.INFO, "Syncing from=3000000, to=3999999")
+                )
+                assert 1 == caplog.record_tuples.count(
+                    (LOG.name, logging.INFO, "Syncing from=4000000, to=4999999")
+                )
+                assert 1 == caplog.record_tuples.count(
+                    (LOG.name, logging.INFO, "Syncing from=5000000, to=5999999")
+                )
+                assert 1 == caplog.record_tuples.count(
+                    (LOG.name, logging.INFO, "Syncing from=6000000, to=6999999")
+                )
+                assert 1 == caplog.record_tuples.count(
+                    (LOG.name, logging.INFO, "Syncing from=7000000, to=7999999")
+                )
+                assert 1 == caplog.record_tuples.count(
+                    (LOG.name, logging.INFO, "Syncing from=8000000, to=8999999")
+                )
+                assert 1 == caplog.record_tuples.count(
+                    (LOG.name, logging.INFO, "Syncing from=9000000, to=9999999")
+                )
+
+        with mock.patch(
+            "web3.eth.async_eth.AsyncEth.block_number", block_number_mock()
+        ):
+            with mock.patch.object(
+                Processor, "_Processor__sync_register", return_value=mock_lib
+            ) as __sync_register_mock:
+                __sync_register_mock.return_value = None
+                await processor.process()
+
+                # Then processor does not call "__sync_register" method.
+                assert __sync_register_mock.call_count == 0
+
+    # <Normal_3_2>
+    # When processed block_number is 9,999,999 and current block number is 19,999,999,
+    # then processor should call "__sync_register" method 10 times.
+    async def test_normal_3_2(
+        self,
+        processor: Processor,
+        shared_contract,
+        async_session: AsyncSession,
+        block_number: None,
+        caplog: pytest.LogCaptureFixture,
+    ):
+        token_list_contract = shared_contract["TokenList"]
+
+        config.TOKEN_LIST_CONTRACT_ADDRESS = token_list_contract["address"]
+
+        mock_lib = MagicMock()
+
+        latest_block_number = 10000000 - 1
+        _token_list_block_number = IDXTokenListBlockNumber()
+        _token_list_block_number.latest_block_number = latest_block_number
+        _token_list_block_number.contract_address = token_list_contract["address"]
+        async_session.add(_token_list_block_number)
+        await async_session.commit()
+
+        current_block_number = 20000000 - 1
+        block_number_mock = AsyncMock()
+        block_number_mock.return_value = current_block_number
+
+        # Run target process
+        with mock.patch(
+            "web3.eth.async_eth.AsyncEth.block_number", block_number_mock()
+        ):
+            with mock.patch.object(
+                Processor, "_Processor__sync_register", return_value=mock_lib
+            ) as __sync_register_mock:
+                __sync_register_mock.return_value = None
+                await processor.process()
+
+                async_session.expunge_all()
+                # Then processor calls "__sync_register" method 10 times.
+                assert __sync_register_mock.call_count == 10
+
+                idx_token_list_block_number: IDXTokenListBlockNumber = (
+                    await async_session.scalars(
+                        select(IDXTokenListBlockNumber).limit(1)
+                    )
+                ).first()
+                assert (
+                    idx_token_list_block_number.latest_block_number
+                    == current_block_number
+                )
+
+                assert 1 == caplog.record_tuples.count(
+                    (LOG.name, logging.INFO, "Syncing from=10000000, to=10999999")
+                )
+                assert 1 == caplog.record_tuples.count(
+                    (LOG.name, logging.INFO, "Syncing from=11000000, to=11999999")
+                )
+                assert 1 == caplog.record_tuples.count(
+                    (LOG.name, logging.INFO, "Syncing from=12000000, to=12999999")
+                )
+                assert 1 == caplog.record_tuples.count(
+                    (LOG.name, logging.INFO, "Syncing from=13000000, to=13999999")
+                )
+                assert 1 == caplog.record_tuples.count(
+                    (LOG.name, logging.INFO, "Syncing from=14000000, to=14999999")
+                )
+                assert 1 == caplog.record_tuples.count(
+                    (LOG.name, logging.INFO, "Syncing from=15000000, to=15999999")
+                )
+                assert 1 == caplog.record_tuples.count(
+                    (LOG.name, logging.INFO, "Syncing from=16000000, to=16999999")
+                )
+                assert 1 == caplog.record_tuples.count(
+                    (LOG.name, logging.INFO, "Syncing from=17000000, to=17999999")
+                )
+                assert 1 == caplog.record_tuples.count(
+                    (LOG.name, logging.INFO, "Syncing from=18000000, to=18999999")
+                )
+                assert 1 == caplog.record_tuples.count(
+                    (LOG.name, logging.INFO, "Syncing from=19000000, to=19999999")
+                )
+
+        with mock.patch(
+            "web3.eth.async_eth.AsyncEth.block_number", block_number_mock()
+        ):
+            with mock.patch.object(
+                Processor, "_Processor__sync_register", return_value=mock_lib
+            ) as __sync_register_mock:
+                __sync_register_mock.return_value = None
+                await processor.process()
+
+                # Then processor does not call "__sync_register" method.
+                assert __sync_register_mock.call_count == 0
+
+    # <Normal_3_3>
+    # When processed block_number is 19,999,999 and current block number is 19,999,999,
+    # then processor should not call "__sync_register" method.
+    async def test_normal_3_3(
+        self,
+        processor: Processor,
+        shared_contract,
+        async_session: AsyncSession,
+        block_number: None,
+        caplog: pytest.LogCaptureFixture,
+    ):
+        token_list_contract = shared_contract["TokenList"]
+
+        config.TOKEN_LIST_CONTRACT_ADDRESS = token_list_contract["address"]
+
+        mock_lib = MagicMock()
+
+        latest_block_number = 20000000 - 1
+        _token_list_block_number = IDXTokenListBlockNumber()
+        _token_list_block_number.latest_block_number = latest_block_number
+        _token_list_block_number.contract_address = token_list_contract["address"]
+        async_session.add(_token_list_block_number)
+        await async_session.commit()
+
+        current_block_number = 20000000 - 1
+        block_number_mock = AsyncMock()
+        block_number_mock.return_value = current_block_number
+
+        # Run target process
+        with mock.patch(
+            "web3.eth.async_eth.AsyncEth.block_number", block_number_mock()
+        ):
+            with mock.patch.object(
+                Processor, "_Processor__sync_register", return_value=mock_lib
+            ) as __sync_register_mock:
+                __sync_register_mock.return_value = None
+                await processor.process()
+
+                # Then processor does not call "__sync_register" method.
+                assert __sync_register_mock.call_count == 0
 
     ###########################################################################
     # Error Case
     ###########################################################################
+    # <Error_1>: ABIEventNotFound occurs in __sync_xx method.
+    # <Error_2_1>: ServiceUnavailable occurs in __sync_xx method.
+    # <Error_2_2>: SQLAlchemyError occurs in "process".
+    # <Error_3>: ServiceUnavailable occurs and is handled in mainloop.
 
-    # <Error_1_1>
-    # API error
-    # Connection error
+    # <Error_1>: ABIEventNotFound occurs in __sync_xx method.
     @mock.patch(
-        "aiohttp.client.ClientSession.get",
-        MagicMock(side_effect=requests.exceptions.ConnectionError),
+        "web3.eth.async_eth.AsyncEth.get_logs",
+        MagicMock(side_effect=ABIEventNotFound()),
     )
-    async def test_error_1_1(self, processor, async_session):
-        # Prepare data
-        _token_list_item = TokenList()
-        _token_list_item.token_address = self.token_address_1
-        _token_list_item.token_template = "ibetBond"
-        _token_list_item.key_manager = ["0000000000000"]
-        _token_list_item.product_type = 1
-        async_session.add(_token_list_item)
-        _token_list_item = TokenList()
-        _token_list_item.token_address = self.token_address_2
-        _token_list_item.token_template = "ibetBond"
-        _token_list_item.key_manager = ["0000000000000"]
-        _token_list_item.product_type = 1
-        async_session.add(_token_list_item)
-        _token_list_item = TokenList()
-        _token_list_item.token_address = self.token_address_3
-        _token_list_item.token_template = "ibetShare"
-        _token_list_item.key_manager = ["1111111111111"]
-        _token_list_item.product_type = 5
-        async_session.add(_token_list_item)
+    async def test_error_1(self, processor: Processor, shared_contract, async_session):
+        # Issue Token
+        token_list_contract = shared_contract["TokenList"]
+        exchange_contract = shared_contract["IbetStraightBondExchange"]
+
+        _token_list_block_number = IDXTokenListBlockNumber()
+        _token_list_block_number.latest_block_number = web3.eth.block_number
+        _token_list_block_number.contract_address = token_list_contract["address"]
+        async_session.add(_token_list_block_number)
         await async_session.commit()
+
+        args = {
+            "name": "テストクーポン",
+            "symbol": "COUPON",
+            "totalSupply": 1000000,
+            "tradableExchange": exchange_contract["address"],
+            "details": "クーポン詳細",
+            "returnDetails": "リターン詳細",
+            "memo": "クーポンメモ欄",
+            "expirationDate": "20191231",
+            "transferable": True,
+            "contactInformation": "問い合わせ先",
+            "privacyPolicy": "プライバシーポリシー",
+        }
+        token = self.issue_token_coupon_with_args(
+            self.issuer, token_list_contract, args
+        )
+        await self.listing_token(token["address"], async_session)
+
+        block_number_current = web3.eth.block_number
+        # Run initial sync
+        await processor.process()
+
+        # Assertion
+        async_session.expunge_all()
+        _token_list = (await async_session.scalars(select(IDXTokenListItem))).all()
+        _token_list_block_number: IDXTokenListBlockNumber = (
+            await async_session.scalars(select(IDXTokenListBlockNumber))
+        ).first()
+        assert len(_token_list) == 0
+        # Latest_block is incremented in "process" process.
+        assert _token_list_block_number.latest_block_number == block_number_current
+
+        token = self.issue_token_coupon_with_args(
+            self.issuer, token_list_contract, args
+        )
+        await self.listing_token(token["address"], async_session)
+
+        block_number_current = web3.eth.block_number
+        # Run target process
+        await processor.process()
 
         # Run target process
         await processor.process()
 
         # Assertion
-        _token_list: Sequence[TokenList] = (
-            await async_session.scalars(
-                select(TokenList).order_by(TokenList.token_address)
-            )
-        ).all()
-        assert len(_token_list) == 3
-
-    # <Error_1_2>
-    # API error
-    # not succeed api
-    @mock.patch("aiohttp.client.ClientSession.get")
-    async def test_error_1_2(self, mock_get, processor, async_session):
-        # Prepare data
-        _token_list_item = TokenList()
-        _token_list_item.token_address = self.token_address_1
-        _token_list_item.token_template = "ibetBond"
-        _token_list_item.key_manager = ["0000000000000"]
-        _token_list_item.product_type = 1
-        async_session.add(_token_list_item)
-        _token_list_item = TokenList()
-        _token_list_item.token_address = self.token_address_2
-        _token_list_item.token_template = "ibetBond"
-        _token_list_item.key_manager = ["0000000000000"]
-        _token_list_item.product_type = 1
-        async_session.add(_token_list_item)
-        _token_list_item = TokenList()
-        _token_list_item.token_address = self.token_address_3
-        _token_list_item.token_template = "ibetShare"
-        _token_list_item.key_manager = ["1111111111111"]
-        _token_list_item.product_type = 5
-        async_session.add(_token_list_item)
-        await async_session.commit()
-
-        # Mock
-        mock_get.side_effect = [MockResponse([], 400)]
-
-        # Run target process
-        await processor.process()
-
+        async_session.expunge_all()
+        await async_session.rollback()
         # Assertion
-        _token_list: Sequence[TokenList] = (
-            await async_session.scalars(
-                select(TokenList).order_by(TokenList.token_address)
-            )
-        ).all()
-        assert len(_token_list) == 3
+        _token_list = (await async_session.scalars(select(IDXTokenListItem))).all()
+        _token_list_block_number: IDXTokenListBlockNumber = (
+            await async_session.scalars(select(IDXTokenListBlockNumber).limit(1))
+        ).first()
+        assert len(_token_list) == 0
+        # Latest_block is incremented in "process" process.
+        assert _token_list_block_number.latest_block_number == block_number_current
 
-    # <Error_2>
-    # not decode response
-    @mock.patch(
-        "aiohttp.client.ClientSession.get",
-        MagicMock(side_effect=json.decoder.JSONDecodeError),
-    )
-    async def test_error_2(self, processor, async_session):
-        # Prepare data
-        _token_list_item = TokenList()
-        _token_list_item.token_address = self.token_address_1
-        _token_list_item.token_template = "ibetBond"
-        _token_list_item.key_manager = ["0000000000000"]
-        _token_list_item.product_type = 1
-        async_session.add(_token_list_item)
-        _token_list_item = TokenList()
-        _token_list_item.token_address = self.token_address_2
-        _token_list_item.token_template = "ibetBond"
-        _token_list_item.key_manager = ["0000000000000"]
-        _token_list_item.product_type = 1
-        async_session.add(_token_list_item)
-        _token_list_item = TokenList()
-        _token_list_item.token_address = self.token_address_3
-        _token_list_item.token_template = "ibetShare"
-        _token_list_item.key_manager = ["1111111111111"]
-        _token_list_item.product_type = 5
-        async_session.add(_token_list_item)
+    # <Error_2_1>: ServiceUnavailable occurs in __sync_xx method.
+    async def test_error_2_1(
+        self, processor: Processor, shared_contract, async_session
+    ):
+        # Issue Token
+        token_list_contract = shared_contract["TokenList"]
+        exchange_contract = shared_contract["IbetStraightBondExchange"]
+
+        _token_list_block_number = IDXTokenListBlockNumber()
+        _token_list_block_number.latest_block_number = web3.eth.block_number
+        _token_list_block_number.contract_address = token_list_contract["address"]
+        async_session.add(_token_list_block_number)
         await async_session.commit()
 
-        # Run target process
-        await processor.process()
+        args = {
+            "name": "テストクーポン",
+            "symbol": "COUPON",
+            "totalSupply": 1000000,
+            "tradableExchange": exchange_contract["address"],
+            "details": "クーポン詳細",
+            "returnDetails": "リターン詳細",
+            "memo": "クーポンメモ欄",
+            "expirationDate": "20191231",
+            "transferable": True,
+            "contactInformation": "問い合わせ先",
+            "privacyPolicy": "プライバシーポリシー",
+        }
+        token = self.issue_token_coupon_with_args(
+            self.issuer, token_list_contract, args
+        )
+        await self.listing_token(token["address"], async_session)
 
+        _token_list_block_number_bf: IDXTokenListBlockNumber = (
+            await async_session.scalars(select(IDXTokenListBlockNumber).limit(1))
+        ).first()
+        # Expect that process() raises ServiceUnavailable.
+        with (
+            mock.patch(
+                "web3.AsyncWeb3.AsyncHTTPProvider.make_request",
+                MagicMock(side_effect=ServiceUnavailable()),
+            ),
+            pytest.raises(ServiceUnavailable),
+        ):
+            await processor.process()
+
+        await async_session.rollback()
         # Assertion
-        _token_list: Sequence[TokenList] = (
-            await async_session.scalars(
-                select(TokenList).order_by(TokenList.token_address)
-            )
-        ).all()
-        assert len(_token_list) == 3
+        _token_list = (await async_session.scalars(select(IDXTokenListItem))).all()
+        _token_list_block_number_af: IDXTokenListBlockNumber = (
+            await async_session.scalars(select(IDXTokenListBlockNumber).limit(1))
+        ).first()
+        assert len(_token_list) == 0
+        assert (
+            _token_list_block_number_bf.latest_block_number
+            == _token_list_block_number_af.latest_block_number
+        )
 
-    # <Error_3>
-    # other error
-    @mock.patch("aiohttp.client.ClientSession.get")
-    async def test_error_3(self, mock_get, processor, async_session):
-        # Prepare data
-        _token_list_item = TokenList()
-        _token_list_item.token_address = self.token_address_1
-        _token_list_item.token_template = "ibetBond"
-        _token_list_item.key_manager = ["0000000000000"]
-        _token_list_item.product_type = 1
-        async_session.add(_token_list_item)
-        _token_list_item = TokenList()
-        _token_list_item.token_address = self.token_address_2
-        _token_list_item.token_template = "ibetBond"
-        _token_list_item.key_manager = ["0000000000000"]
-        _token_list_item.product_type = 1
-        async_session.add(_token_list_item)
-        _token_list_item = TokenList()
-        _token_list_item.token_address = self.token_address_3
-        _token_list_item.token_template = "ibetShare"
-        _token_list_item.key_manager = ["1111111111111"]
-        _token_list_item.product_type = 5
-        async_session.add(_token_list_item)
-        await async_session.commit()
+        token = self.issue_token_coupon_with_args(
+            self.issuer, token_list_contract, args
+        )
+        await self.listing_token(token["address"], async_session)
 
-        # Mock
-        mock_get.side_effect = [
-            MockResponse(
-                [
-                    {
-                        "token_address": self.token_address_4,
-                        "token_template": "ibetShare",
-                        "key_manager": ["0000000000000", "1111111111111"],
-                        "product_type": 1,
-                    },
-                    "aaaaaaaa",  # not dict
-                ]
-            )
-        ]
+        await async_session.rollback()
+        _token_list_block_number_bf: IDXTokenListBlockNumber = (
+            await async_session.scalars(select(IDXTokenListBlockNumber).limit(1))
+        ).first()
 
-        # Run target process
-        with pytest.raises(Exception):
+        # Expect that process() raises ServiceUnavailable.
+        with (
+            mock.patch(
+                "web3.AsyncWeb3.AsyncHTTPProvider.make_request",
+                MagicMock(side_effect=ServiceUnavailable()),
+            ),
+            pytest.raises(ServiceUnavailable),
+        ):
             await processor.process()
 
         # Assertion
-        _token_list: Sequence[TokenList] = (
-            await async_session.scalars(
-                select(TokenList).order_by(TokenList.token_address)
-            )
-        ).all()
-        assert len(_token_list) == 3
+        await async_session.rollback()
+        _token_list = (await async_session.scalars(select(IDXTokenListItem))).all()
+        _token_list_block_number_af: IDXTokenListBlockNumber = (
+            await async_session.scalars(select(IDXTokenListBlockNumber).limit(1))
+        ).first()
+        assert len(_token_list) == 0
+        assert (
+            _token_list_block_number_bf.latest_block_number
+            == _token_list_block_number_af.latest_block_number
+        )
+
+    # <Error_2_2>: SQLAlchemyError occurs in "process".
+    async def test_error_2_2(
+        self, processor: Processor, shared_contract, async_session
+    ):
+        # Issue Token
+        token_list_contract = shared_contract["TokenList"]
+        exchange_contract = shared_contract["IbetStraightBondExchange"]
+
+        _token_list_block_number = IDXTokenListBlockNumber()
+        _token_list_block_number.latest_block_number = web3.eth.block_number
+        _token_list_block_number.contract_address = token_list_contract["address"]
+        async_session.add(_token_list_block_number)
+        await async_session.commit()
+
+        args = {
+            "name": "テストクーポン",
+            "symbol": "COUPON",
+            "totalSupply": 1000000,
+            "tradableExchange": exchange_contract["address"],
+            "details": "クーポン詳細",
+            "returnDetails": "リターン詳細",
+            "memo": "クーポンメモ欄",
+            "expirationDate": "20191231",
+            "transferable": True,
+            "contactInformation": "問い合わせ先",
+            "privacyPolicy": "プライバシーポリシー",
+        }
+        token = self.issue_token_coupon_with_args(
+            self.issuer, token_list_contract, args
+        )
+        await self.listing_token(token["address"], async_session)
+
+        # Expect that process() raises SQLAlchemyError.
+        with (
+            mock.patch.object(Session, "commit", side_effect=SQLAlchemyError()),
+            pytest.raises(SQLAlchemyError),
+        ):
+            await processor.process()
+
+        # Assertion
+        _token_list = (await async_session.scalars(select(IDXTokenListItem))).all()
+        _token_list_block_number_af: IDXTokenListBlockNumber = (
+            await async_session.scalars(select(IDXTokenListBlockNumber).limit(1))
+        ).first()
+        assert len(_token_list) == 0
+        assert (
+            _token_list_block_number_af.latest_block_number
+            == _token_list_block_number.latest_block_number
+        )
+
+        token = self.issue_token_coupon_with_args(
+            self.issuer, token_list_contract, args
+        )
+        await self.listing_token(token["address"], async_session)
+
+        await async_session.rollback()
+
+        # Expect that process() raises SQLAlchemyError.
+        with (
+            mock.patch.object(Session, "commit", side_effect=SQLAlchemyError()),
+            pytest.raises(SQLAlchemyError),
+        ):
+            await processor.process()
+
+        # Assertion
+        await async_session.rollback()
+        _token_list = (await async_session.scalars(select(IDXTokenListItem))).all()
+        _token_list_block_number_af: IDXTokenListBlockNumber = (
+            await async_session.scalars(select(IDXTokenListBlockNumber).limit(1))
+        ).first()
+        assert len(_token_list) == 0
+        assert (
+            _token_list_block_number_af.latest_block_number
+            == _token_list_block_number.latest_block_number
+        )
+
+    # <Error_3>: ServiceUnavailable occurs and is handled in mainloop.
+    async def test_error_3(self, main_func, shared_contract, async_session, caplog):
+        # Issue Token
+        token_list_contract = shared_contract["TokenList"]
+        exchange_contract = shared_contract["IbetStraightBondExchange"]
+
+        args = {
+            "name": "テストクーポン",
+            "symbol": "COUPON",
+            "totalSupply": 1000000,
+            "tradableExchange": exchange_contract["address"],
+            "details": "クーポン詳細",
+            "returnDetails": "リターン詳細",
+            "memo": "クーポンメモ欄",
+            "expirationDate": "20191231",
+            "transferable": True,
+            "contactInformation": "問い合わせ先",
+            "privacyPolicy": "プライバシーポリシー",
+        }
+        token = self.issue_token_coupon_with_args(
+            self.issuer, token_list_contract, args
+        )
+        await self.listing_token(token["address"], async_session)
+
+        # Mocking time.sleep to break mainloop
+        asyncio_mock = MagicMock(wraps=asyncio)
+        asyncio_mock.sleep.side_effect = [TypeError()]
+
+        # Run mainloop once and fail with web3 utils error
+        with (
+            mock.patch("batch.indexer_Token_List.asyncio", asyncio_mock),
+            mock.patch(
+                "web3.AsyncWeb3.AsyncHTTPProvider.make_request",
+                MagicMock(side_effect=ServiceUnavailable()),
+            ),
+            pytest.raises(TypeError),
+        ):
+            # Expect that process() raises ServiceUnavailable and handled in mainloop.
+            await main_func()
+
+        assert 1 == caplog.record_tuples.count(
+            (LOG.name, logging.INFO, "Service started successfully")
+        )
+        assert 1 == caplog.record_tuples.count(
+            (LOG.name, 25, "An external service was unavailable")
+        )
+        caplog.clear()
